don't use AgentMCPResource in mcp.ts
